refactor HttpHeadersPolicy
