+++ src/main/java/dev/vality/wachter/config/http/HttpHeadersPolicy.java
@@ -1,0 +1,49 @@
+package dev.vality.wachter.config.http;
+
+import org.apache.hc.core5.http.HttpHeaders;
+import org.springframework.stereotype.Component;
+
+import java.util.Collections;
+import java.util.List;
+import java.util.Locale;
+import java.util.Set;
+
+@Component
+public class HttpHeadersPolicy {
+
+    /**
+     * Headers that must be excluded from incoming requests before building outbound request.
+     * These headers should be generated by the HTTP client itself based on the request body and target host.
+     */
+    private static final Set<String> EXCLUDED_REQUEST_HEADERS = Set.of(
+            HttpHeaders.HOST.toLowerCase(Locale.ROOT),
+            HttpHeaders.CONTENT_LENGTH.toLowerCase(Locale.ROOT),
+            HttpHeaders.TRANSFER_ENCODING.toLowerCase(Locale.ROOT)
+    );
+
+    /**
+     * Hop-by-hop headers that must be removed before forwarding the request to upstream.
+     * These headers are connection-specific and should not be forwarded by proxies (RFC 7230).
+     */
+    private static final Set<String> OUTBOUND_SANITIZED_HEADERS = Set.of(
+            HttpHeaders.CONNECTION.toLowerCase(Locale.ROOT),
+            HttpHeaders.TE.toLowerCase(Locale.ROOT),
+            HttpHeaders.PROXY_AUTHENTICATE.toLowerCase(Locale.ROOT),
+            HttpHeaders.PROXY_AUTHORIZATION.toLowerCase(Locale.ROOT),
+            HttpHeaders.PROXY_CONNECTION.toLowerCase(Locale.ROOT),
+            HttpHeaders.TRAILER.toLowerCase(Locale.ROOT),
+            HttpHeaders.UPGRADE.toLowerCase(Locale.ROOT)
+    );
+
+    public boolean shouldExcludeFromRequest(String headerName) {
+        return headerName != null && EXCLUDED_REQUEST_HEADERS.contains(headerName.toLowerCase(Locale.ROOT));
+    }
+
+    public Set<String> getOutboundSanitizedHeaders() {
+        return OUTBOUND_SANITIZED_HEADERS;
+    }
+
+    public List<String> getCorsAllowedHeaders() {
+        return Collections.singletonList("*");
+    }
+}
